move ZMQ-specific code to utils
We want to use ZMQ to send both register reads/writes to the model as well
as perform standard memory accesses. Move it to a shared place so that
both the board implementation and the CSR memory map can use it.
